+++ pkg/component/types.go
@@ -1,8 +1,6 @@
 package component
 
 import (
-	"github.com/openshift/odo/pkg/config"
-	"github.com/openshift/odo/pkg/occlient"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
@@ -36,13 +34,4 @@ type ComponentStatus struct {
 	Active           bool                `json:"active"`
 	LinkedComponents map[string][]string `json:"linkedComponents,omitempty"`
 	LinkedServices   []string            `json:"linkedServices,omitempty"`
-}
-
-// ComponentOptions are generic options that are used between commands such as watch and push
-type ComponentOptions struct {
-	SourceType       config.SrcType
-	SourcePath       string
-	LocalConfig      *config.LocalConfigInfo
-	ComponentContext string
-	Client           *occlient.Client
 }
